+++ R/tadaa_pairwise_t.R
@@ -24,6 +24,7 @@
 #' @return A \code{data.frame} with columns \code{term}, \code{comparison} and \code{adj.p.value}.
 #' @export
 #' @family Tadaa-functions
+#' @seealso \code{\link{tadaa_pairwise_tukey}}, \code{\link{tadaa_pairwise_gh}}
 #' @note The adjustment method is applied within each \code{term}, meaning that the number of
 #' pairwise t-tests counted for the adjustment is only equal to the number of rows
 #' per \code{term} of the output.